Update README.md
# Kickstarting with Excel

### The purpose of this project is to visualize campaign outcomes based on the their launching dates and the funding goals

## The analysis was performed as I created pivot tables and then using pivot charts to visualize the trends. One challenge that I faced was having the months on the pivot table for the theatre outcomes based on the land date. Therefore, I created another columns where it shows the months of the launching dates. --- [Kickstarter_Challenge](path/to/fKickstarter.xlxs).
### ![Theater_Outcomes_Based_on_Launching_Dates] (path/to/Theater_Outcomes_Based_on_Launcing_Dates.png)
### ![Outcomes_vs_Goals] (path/to/Outcomes_vs_Goals.png)

### One challenge that I faced was having the months on the pivot table for the theatre outcomes based on the land date. Therefore, I created another columns where it shows the months of the launching dates. 

## Results
The outcomes based on the launching dates for the theater subcategory are overall successful, as there is more successful outcomes compared too failed and canceled. However, the trending of this outcomes are not consistent as each month are drastic drops and up. Some limitations is that there is no comparison with different years to determine a more complete conclusion. Some other graph to be used can be column graphs.
We can see that the outcomes based on Goal had a successful outcomes when the goal is less the $15,000 and then the outcomes failed when the goal is higher. There is an even outcome around 50% successful and failed on the ranges $15,000-$20,000 , $35,000 and $40,000-45,000. We can also conclude that successful outcomes are dropping as the goals are high amount. Also, we can see that there was no canceled outcomes based on different goals.